reworked Iroha_CommandError in proto_command_executor
