+++ irohad/ametsuchi/impl/proto_command_executor.cpp
@@ -6,25 +6,19 @@
 #include "ametsuchi/impl/proto_command_executor.h"
 
 #include "ametsuchi/command_executor.hpp"
+#include "ametsuchi/impl/common_c_types_helpers.hpp"
 #include "backend/protobuf/commands/proto_command.hpp"
 #include "validators/field_validator.hpp"
 #include "validators/protobuf/proto_command_validator.hpp"
 #include "validators/transaction_validator.hpp"
 #include "validators/validators_common.hpp"
 
-namespace {
-  char *clone(const std::string &string) {
-    char *cstr = new char[string.length() + 1];
-    strcpy(cstr, string.c_str());
-    return cstr;
-  }
-}  // namespace
-
 Iroha_CommandError Iroha_ProtoCommandExecutorExecute(void *executor,
                                                      void *data,
                                                      int size,
                                                      char *account_id) {
   Iroha_CommandError result{};
+  iroha::clearCharBuffer(result.command_name);
 
   iroha::protocol::Command protocol_command;
   if (!protocol_command.ParseFromArray(data, size)) {
@@ -36,7 +30,7 @@ Iroha_CommandError Iroha_ProtoCommandExecutorExecute(void *executor,
           shared_model::validation::ProtoCommandValidator().validate(
               protocol_command)) {
     result.error_code = 200;
-    result.error_extra = clone(maybe_error.value().toString());
+    iroha::toCharBuffer(result.error_extra, maybe_error.value().toString());
     return result;
   }
 
@@ -50,7 +44,7 @@ Iroha_CommandError Iroha_ProtoCommandExecutorExecute(void *executor,
 
   if (maybe_error) {
     result.error_code = 300;
-    result.error_extra = clone(maybe_error.value().toString());
+    iroha::toCharBuffer(result.error_extra, maybe_error.value().toString());
     return result;
   }
 
@@ -59,7 +53,7 @@ Iroha_CommandError Iroha_ProtoCommandExecutorExecute(void *executor,
       .match([](const auto &) {},
              [&result](const auto &error) {
                result.error_code = error.error.error_code;
-               result.error_extra = clone(error.error.error_extra);
+               iroha::toCharBuffer(result.error_extra, error.error.error_extra);
              });
   return result;
 }
